requirements.txt
streamlit
openai-whisper
ffmpeg
langchain
langchain_google_genai
python-dotenv
zyphra
